lm-agent/requirements.txt
httpx>=0.27
rich>=13.0
prompt-toolkit>=3.0
PyYAML>=6.0
